fixes per pr comments on god mode
