+++ packages/commonwealth/client/scripts/views/pages/AdminPanel/index.ts
@@ -1,0 +1,2 @@
+import AdminPanelPage from './AdminPanelPage';
+export default AdminPanelPage;
