+++ packages/commonwealth/server/routes/createChainNode.ts
@@ -1,12 +1,6 @@
 import { AppError } from 'common-common/src/errors';
-import type { NextFunction, Request, Response } from 'express';
 import type { DB } from '../models';
-import {
-  TypedRequestBody,
-  TypedRequestQuery,
-  TypedResponse,
-  success,
-} from '../types';
+import { TypedRequestBody, TypedResponse, success } from '../types';
 import { BalanceType } from '../../../common-common/src/types';
 
 export const Errors = {
